Add NavigationSection component
Added the NavigationSection and NavigationItem component to handle navigation in the application. In this component, a "change color on hover" feature is implemented for better user experience. This color transition corresponds with a timeout delay; hence the use of useEffect here is to handle cleanup and prevent the change color effect from carrying out in case the user leaves the navigation item before the timeout expires. This will help improve navigation in the application, allowing users to move around different sections conveniently and enjoyably.