Add new contact page
A new contact page was created using the 'DefaultPageLayout' component. This was done to provide users with a simple and efficient way to get in contact with the website owner. The page includes a form for communication and a list of contacts. Also, a message indicating that the website owner is seeking a job was added to increase potential employment opportunities.