Refactor UI components and update session handling
This commit contains updates primarily aimed at improving the website's usability and the code's organization.

For the ViewTransitionModal component, a conditional has been added to ensure the window object exists before trying to access the session storage. This prevents potential server-side rendering issues which could lead to a crash, as the window object is not available on the server-side.

The ANIMATE_TEXT constant has been tweaked to use a hyphen instead of a space, thereby enhancing readability.

The way attributes spread over the NavigationItem component has been adjusted, changing the order to distribute the route properties after the key is set. This change guards against potential overwrites of the key property if it were in the route object.

Two new transition constants have been added in the transitions.ts file: SLIDE_RIGHT_TRANSITION and OPACITY_TRANSITION. This explicit declaration supports the use of these transitions across multiple components increasing reusability and maintainability of the code.

Finally, the icons used in the routes.tsx and GoHomeLink.tsx components have been moved into a new 'navigation' directory. This better organizes these components and separates their concerns from the other icons not specifically related to navigation.