Refactor Dialog component for improved UX
In this commit, the code base for the Dialog component was refactored. The primary reason for this change was to improve user experience.

Previously, we were using a simple Transition component for dialog transitions. Now, using the AnimatePresence and motion modules from the framer-motion library, we have made the animation more lively and engaging and so, enhancing the UX.

We also added escape key functionality to close the dialog to improve accessibility. We also started using useRef hook to ensure that the focus is kept within the dialog when it's open, again, a vital part of ensuring our applications accessibility.