"Add tech stack icons for listing in portfolio"
This commit adds various SVG components that represent the different technologies used in various projects in the portfolio. Each icon acts as a visual representation of a particular technology, such as Node.js, GraphQL, Express, TypeScript, etc. They have been added to enhance the UI of the portfolio, providing a clear and visually appealing way to display the tech stack used in the projects.