Add StackSection component to display tech stack
This commit introduces a new component, StackSection, in 'src/components'. This component takes a list of tech stack items as props and presents them in a stylistic manner. The component uses animations from 'framer-motion' to improve the user experience when the items appear. The addition is a part of improving the overall aesthetics and user experience of the application.