Add new components to About page
New components were added to the About page of the application. React was imported, followed by EducationSlider, TextAnimation and about constants from the config directory. Adjustments were made to the DefaultPageLayout alongside the addition of TextAnimation and EducationSlider. This is done to expand on the educational information present and to create a more dynamic and engaging About page layout.