Update Projects page with dynamic project data
Included functionality to render a list of dynamically loaded project cards in the Projects page. This was done by importing a getProjectsData function and PROJECTS_DIR_PATH constant from lib and config folders respectively. The Project card data is fetched asynchronously in the Projects function, and then mapped to ProjectCard components that are displayed in a flex container. This change is intended to enhance the scalability and maintenance efficiency of the Projects page by allowing easy addition and modification of project cards.