Replace useAnimatedRouter with useRouter
Removed useAnimatedRouter custom hook and replaced it with the default useRouter from "next/navigation" library. The change was made in navigationItem.tsx file in NavigationSection component. The reason behind this change was to simplify the routing mechanism and resolve potential animation issues which were previously handled by useAnimatedRouter hook. Now, on click of navigation item, setTheme and push actions are called to navigate to the new route.