Update embedIds, order and titles in project files
In this commit, all the changes revolve around improving the organisation of project information. The updated `embedId`s point to more relevant sources that better represent the projects. Furthermore, the 'Authentication App' project and 'pangeayt-2-calc' project's order was adjusted for better visibility and preference. Lastly, 'pangeayt-2-calc' project's title was tweaked for consistency. These changes will help make the visual representation of the projects more accurate and the site easier to navigate.