Add form validation and email sending functionality
This commit introduces a mechanism to validate a user's form inputs as well as sending the data via email. A `validate-form.tsx` file has been added in the utils directory to validate the name, email, and message fields of our contact form. This will throw an error if any required field is left empty or if the email is invalid.

Environment variable check is added in the `send-email.ts` to ensure all necessary environment variables are present for sending an email. We are also catching and returning errors in case anything wrong happens.

Also, a ResultDialog has been added to the ContactForm to feedback the user about an error or success message after the form submission. Successful form capture leads to sending an email, and after that, triggering a dialog to inform user about the operation result.

This ensures a better user experience as users are now informed whether their input is acceptable and whether their message has been successfully sent.