Add ResultDialog component
The `ResultDialog` component has been introduced for displaying messages to the user. It can be used to display feedback, errors or simple informational messages. It takes in `isOpen`, `message`, and `handleClose` props. `isOpen` controls the visibility of the dialog, `message` is the text to be displayed and `handleClose` is a function to close the dialog.