Update UI responsiveness and styling
Updated several CSS classes to improve the responsiveness and visual aesthetics across the application. This has involved modifying the ProjectCardAnimation component to use a larger delay class, adjusting the size and position styles in the layout and page components to make the interface fit any screen more naturally and updating the font weight of the project page titles for better legibility. Similarly, resized text and wrapped content in the resume dialog download section to prevent overflow and ensure it fits within the dialog box regardless of screen size or language selected.