Refactor page layout and add fadeIn animation
This commit removes the use of PageWrapper from the default page layout and home page, and directly nests the main content inside ThemeProvider. This is to simplify the component tree and improve readability.

Additionally, fadeIn animation is added to both the default page layout and home page, resulting in a smoother transition when the components are rendered. This change also includes the adjusted fadeIn animation keyframe in the TailwindCSS configuration.