Add React Testing Library icon component
A new icon component 'ReactTestingLibrary' was added to the 'icons/stack' directory. This component serves as an illustrative icon for React Testing Library technology. It was also exported in the 'index.ts' of the same directory to make it easily accessible in other parts of the project. This addition supports user interface enhancements, specifically in areas where we would want to represent technologies used with their respective icons.