Add handleClick function for navigation items
A handleClick function was added for better code readability and reusability. This function, which is called during the onClick event of a navigation item, is responsible for setting the theme and navigation. The function was also bound to an onKeyDown event for accessibility purposes, allowing theme setting and navigation to be triggered by hitting the Enter key.