Add unit tests for EmailDialog component
A new test file, EmailDialog.test.tsx, has been added to conduct multiple unit tests on the EmailDialog component. These test different functionalities such as rendering based on the isOpen flag, a loading SVG based on the isLoading flag, correct running behaviour of the close method when 'Esc' is pressed, and proper rendering of messages. These tests are to ensure that our component behaves as expected and all possible edge cases are covered.