Add unit tests for LanguageSwitch component
Tests have been added for the LanguageSwitch component. These tests are aimed to verify the component is rendering correctly and its callback (setIsEnglish) is fired as expected when interacting with it. The tests make sure the callback is called when "switch" element is clicked or when Enter key is pressed while it's in focus.